agent/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
blinker==1.8.2
boilerpy3==1.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langfuse==2.36.2
langfuse-haystack==0.2.0
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-plain==1.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
networkx==3.3
numpy==1.26.4
openai==1.30.5
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pgvector==0.2.5
pgvector-haystack==0.3.0
posthog==3.5.0
psycopg==3.1.19
psycopg-binary==3.1.19
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==8.3.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
